Move CRTP BodyTransform next to the only class that uses it
